Autogen/IT-Ticket/requirements.txt
python-dotenv
streamlit
pyautogen
openai
chromadb
pandas
google-search-results